**** INTER ****
- Corrijo rutinas de grabado y lectura de eeprom.
- Actualizo test eeprom.
- Agrego comando de lectura de memoria 'hist'.
- Agrego comando de monitoreo de variables 'mon' y 'stop' para quitarlo.
- Agrego primer soft timer para soporte de comando 'mon'.

+++ tests/mockFiles/inc/api_TYPES.h
@@ -11,8 +11,9 @@
 //#ifdef	TEST_MEMORIA
 
 typedef unsigned char uint8_t;
-typedef unsigned long int uint32_t;
-typedef unsigned short int uint16_t;
+typedef unsigned short uint16_t;
+typedef unsigned int   uint32_t;
+
 //#endif
